Update efficiency_fit.py
Need to change when it plots.  Currently plotting in build_eff_plots.
open_files should be done by the user because their files will be different obviously. I put opening the files in main for that reason.

"... = check_calibration(input_data??)" and "axes = plot_fit_results([(ind_fit_result, ind_plot_axes)], ax=ax)" need to be done if that is still in the plan.